refactor: reorganize into a module structure
